requirements.txt
biopython
chromadb
sentence-transformers
requests